Making RsaPaddingScheme an inner class of RawRsaKeyringBuilder
