+++ src/test/java/com/amazonaws/crypto/examples/RawRsaKeyringEncryptExampleTest.java
@@ -16,8 +16,8 @@ package com.amazonaws.crypto.examples;
 import com.amazonaws.encryptionsdk.AwsCrypto;
 import com.amazonaws.encryptionsdk.DecryptRequest;
 import com.amazonaws.encryptionsdk.keyrings.Keyring;
+import com.amazonaws.encryptionsdk.keyrings.RawRsaKeyringBuilder.RsaPaddingScheme;
 import com.amazonaws.encryptionsdk.keyrings.StandardKeyrings;
-import com.amazonaws.encryptionsdk.model.RsaPaddingScheme;
 import org.junit.jupiter.api.Test;
 
 import java.security.KeyPair;
